(chore) move NewItem into logbooks/
